Move closure_plus and closure_star to closure file
